Create DeletedValidationIT for deletion tests
